+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Update Dropdown:
+        
+    - Add new props: `onDropdownVisibleChange`, `autoFocus`.
+
+- Add new component `Select`    
+    
 ## [0.38.0] - 2021-06-28
 
 - Update PopupNotificationList: